+++ src/apify_client/_http_client.py
@@ -12,10 +12,10 @@ from typing import TYPE_CHECKING, Any
 import httpx
 from apify_shared.utils import ignore_docs, is_content_type_json, is_content_type_text, is_content_type_xml
 
-from apify_client._errors import ApifyApiError, InvalidResponseBodyError, is_retryable_error
 from apify_client._logging import log_context, logger_name
 from apify_client._statistics import Statistics
 from apify_client._utils import retry_with_exp_backoff, retry_with_exp_backoff_async
+from apify_client.errors import ApifyApiError, InvalidResponseBodyError, is_retryable_error
 
 if TYPE_CHECKING:
     from collections.abc import Callable
