+++ src/apify_client/clients/resource_clients/key_value_store.py
@@ -7,9 +7,9 @@ from typing import TYPE_CHECKING, Any
 
 from apify_shared.utils import filter_out_none_values_recursively, ignore_docs, parse_date_fields
 
-from apify_client._errors import ApifyApiError
 from apify_client._utils import catch_not_found_or_throw, encode_key_value_store_record_value, pluck_data
 from apify_client.clients.base import ResourceClient, ResourceClientAsync
+from apify_client.errors import ApifyApiError
 
 if TYPE_CHECKING:
     from collections.abc import AsyncIterator, Iterator
