+++ src/apify_client/clients/resource_clients/log.py
@@ -13,9 +13,9 @@ from typing import TYPE_CHECKING, Any, cast
 
 from apify_shared.utils import ignore_docs
 
-from apify_client._errors import ApifyApiError
 from apify_client._utils import catch_not_found_or_throw
 from apify_client.clients.base import ResourceClient, ResourceClientAsync
+from apify_client.errors import ApifyApiError
 
 if TYPE_CHECKING:
     from collections.abc import AsyncIterator, Iterator
