+++ src/apify_client/clients/resource_clients/request_queue.py
@@ -10,9 +10,9 @@ from typing import TYPE_CHECKING, Any, TypedDict
 from apify_shared.utils import filter_out_none_values_recursively, ignore_docs, parse_date_fields
 from more_itertools import constrained_batches
 
-from apify_client._errors import ApifyApiError
 from apify_client._utils import catch_not_found_or_throw, pluck_data
 from apify_client.clients.base import ResourceClient, ResourceClientAsync
+from apify_client.errors import ApifyApiError
 
 if TYPE_CHECKING:
     from datetime import timedelta
