+++ src/apify_client/clients/resource_clients/schedule.py
@@ -4,9 +4,9 @@ from typing import Any
 
 from apify_shared.utils import filter_out_none_values_recursively, ignore_docs
 
-from apify_client._errors import ApifyApiError
 from apify_client._utils import catch_not_found_or_throw, pluck_data_as_list
 from apify_client.clients.base import ResourceClient, ResourceClientAsync
+from apify_client.errors import ApifyApiError
 
 
 def _get_schedule_representation(
