+++ tests/unit/test_client_request_queue.py
@@ -43,7 +43,7 @@ async def test_batch_not_processed_raises_exception_async(httpserver: HTTPServer
     ]
     rq_client = client.request_queue(request_queue_id='whatever')
 
-    with pytest.raises(apify_client._errors.ApifyApiError):
+    with pytest.raises(apify_client.errors.ApifyApiError):
         await rq_client.batch_add_requests(requests=requests)
 
 
@@ -77,7 +77,7 @@ def test_batch_not_processed_raises_exception_sync(httpserver: HTTPServer) -> No
     ]
     rq_client = client.request_queue(request_queue_id='whatever')
 
-    with pytest.raises(apify_client._errors.ApifyApiError):
+    with pytest.raises(apify_client.errors.ApifyApiError):
         rq_client.batch_add_requests(requests=requests)
